services/policy-engine/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pytesseract==0.3.10
Pillow==10.0.1
PyMuPDF==1.23.8
numpy==1.24.3
python-multipart==0.0.6